# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: Chimera Analytics API
  description: |
    REST API for the Chimera Multi-Agent Analytics System.
    
    The Chimera API provides programmatic access to Zcash analytics capabilities including:
    - Natural language query submission
    - Report retrieval and export
    - Dashboard configuration
    - Alert management
    - Metrics data access
    
    ## Authentication
    
    The API supports two authentication methods:
    - **JWT Bearer Token**: Include in `Authorization: Bearer <token>` header
    - **API Key**: Include in `X-API-Key` header
    
    ## Rate Limiting
    
    API requests are rate limited to 100 requests per minute per user/API key.
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when the rate limit resets
    
  version: 1.0.0
  contact:
    name: Chimera API Support
    email: support@chimera-analytics.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.chimera-analytics.example.com
    description: Production server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/alerts:
    post:
      summary: Create alert rule
      description: Create a new alert rule
      tags:
        - Alerts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleCreate'
      responses:
        '201':
          description: Alert rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'

    get:
      summary: List alert rules
      description: List user's alert rules
      tags:
        - Alerts
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: enabled
          in: query
          schema:
            type: boolean
          description: Filter by enabled status
      responses:
        '200':
          description: Alert rule list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRuleListResponse'
  /api/alerts/{id}:
    get:
      summary: Get alert rule
      description: Retrieve alert rule by ID
      tags:
        - Alerts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert rule retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'

    put:
      summary: Update alert rule
      description: Update an alert rule
      tags:
        - Alerts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleUpdate'
      responses:
        '200':
          description: Alert rule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRule'

    delete:
      summary: Delete alert rule
      description: Delete an alert rule
      tags:
        - Alerts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Alert rule deleted successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
  schemas:
...
    AlertRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        condition:
          $ref: '#/components/schemas/AlertCondition'
        channels:
          type: array
          items:
            type: object
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AlertCondition:
      type: object
      required:
        - metric
        - operator
        - threshold
      properties:
        metric:
          type: string
        operator:
          type: string
          enum: ['>', '<', '==', '!=', 'change_pct']
        threshold:
          type: number
        duration:
          type: integer
        cooldown:
          type: integer
    AlertRuleCreate:
      type: object
      required:
        - name
        - condition
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          $ref: '#/components/schemas/AlertCondition'
        channels:
          type: array
          items:
            type: object
        enabled:
          type: boolean
    AlertRuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          $ref: '#/components/schemas/AlertCondition'
        channels:
          type: array
        enabled:
          type: boolean
    AlertRuleListResponse:
      type: object
      properties:
        rules:
          type: array
          items:
            type: object
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...